refactor: add components folder
