+++ src/modules/reports/forms/create.tsx
@@ -11,19 +11,19 @@ import { MatchBasicDataDto } from '@/modules/matches/types'
 import { PlayerPositionDto } from '@/modules/player-positions/types'
 import { PlayerBasicDataDto } from '@/modules/players/types'
 import { ReportTemplateBasicDataDto } from '@/modules/report-templates/types'
-import { MetaStep } from '@/modules/reports/forms/meta-step'
-import { SkillAssessmentsStep } from '@/modules/reports/forms/skill-assessments-step'
-import { SummaryStep } from '@/modules/reports/forms/summary-step'
 import { TeamBasicDataDto } from '@/modules/teams/types'
 import { useStepper } from '@/utils/hooks/use-stepper'
 
 import { CreateReportDto, ReportType } from '../types'
-import { MatchStep } from './match-step'
-import { PlayerStep } from './player-step'
-import { ReportTypeStep } from './report-type-step'
-import { StatsStep } from './stats-step'
-import { StepActions } from './step-actions'
-import { TemplateStep } from './template-step'
+import { MatchStep } from './components/match-step'
+import { MetaStep } from './components/meta-step'
+import { PlayerStep } from './components/player-step'
+import { ReportTypeStep } from './components/report-type-step'
+import { SkillAssessmentsStep } from './components/skill-assessments-step'
+import { StatsStep } from './components/stats-step'
+import { StepActions } from './components/step-actions'
+import { SummaryStep } from './components/summary-step'
+import { TemplateStep } from './components/template-step'
 import {
   formatCreateReportDto,
   generateReportFormValidationSchema,
